+++ llvm/include/llvm/Transforms/Utils.h
@@ -48,6 +48,12 @@ extern char &InstructionNamerID;
 
 //===----------------------------------------------------------------------===//
 //
+// IRCanonicalizer - Transforms LLVM Modules into canonical form.
+//
+Pass *createIRCanonicalizerPass();
+
+//===----------------------------------------------------------------------===//
+//
 // LowerSwitch - This pass converts SwitchInst instructions into a sequence of
 // chained binary branch instructions.
 //