+++ llvm/docs/ReleaseNotes.rst
@@ -41,7 +41,10 @@ Non-comprehensive list of changes in this release
    for adding a new subsection.
 
 * ...
-
+* Added a new IRCanonicalizer pass which aims to transform LLVM modules into
+  a canonical form by reordering and renaming instructions while preserving the
+  same semantics. The canonicalizer makes it easier to spot semantic differences
+  when diffing two modules which have undergone different passes.
 
 .. NOTE
    If you would like to document a larger change, then you can add a